sets server port, convert metadata id type to int

+++ src/main/java/com/casestudy/busmetadataservice/repository/MetadataRepository.java
@@ -3,5 +3,5 @@ package com.casestudy.busmetadataservice.repository;
 import com.casestudy.busmetadataservice.entity.Metadata;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MetadataRepository extends JpaRepository<Metadata, Long> {
+public interface MetadataRepository extends JpaRepository<Metadata, Integer> {
 }
